Remove excess spaces, make extensions private and put in Geohash.swift
