+++ Sources/Geohash/Geohash.swift
@@ -31,7 +31,7 @@ public enum Geohash {
             .joined(separator: "")
         guard bits.count % 5 == 0 else { return nil }
         // bits = 1101000100101011011111010111100110010110101101101110001
-        
+
         let (lat, lon) = bits.enumerated().reduce(into: ([Character](), [Character]())) {
             if $1.0 % 2 == 0 {
                 $0.1.append($1.1)
@@ -41,24 +41,24 @@ public enum Geohash {
         }
         // lat = [1,1,0,1,0,0,0,1,1,1,1,1,1,1,0,1,0,1,1,0,0,1,1,0,1,0,0]
         // lon = [1,0,0,0,0,1,1,1,0,1,1,0,0,1,1,0,1,0,0,1,1,1,0,1,1,1,0,1]
-        
+
         func combiner(array a: (min: Double, max: Double), value: Character) -> (Double, Double) {
             let mean = (a.min + a.max) / 2
             return value == "1" ? (mean, a.max) : (a.min, mean)
         }
-        
+
         let latRange = lat.reduce((-90.0, 90.0), combiner)
         // latRange = (57.649109959602356, 57.649111300706863)
-        
+
         let lonRange = lon.reduce((-180.0, 180.0), combiner)
         // lonRange = (10.407439023256302, 10.407440364360809)
-        
+
         return (latRange, lonRange)
     }
-    
+
     public static func encode(latitude: Double, longitude: Double, length: Int) -> String {
         // For example: (latitude, longitude) = (57.6491106301546, 10.4074396938086)
-        
+
         func combiner(array a: (min: Double, max: Double, array: [String]), value: Double) -> (Double, Double, [String]) {
             let mean = (a.min + a.max) / 2
             if value < mean {
@@ -73,19 +73,19 @@ public enum Geohash {
 
         let lon = Array(repeating: longitude, count: length * 5).reduce((-180.0, 180.0, [String]()), combiner)
         // lon = (10.407439693808236, 10.407439693808556, [1,0,0,0,0,1,1,1,0,1,1,0,0,1,1,0,1,0,0,1,1,1,0,1,1,1,0,1,0,1,..])
-        
+
         let latlon = lon.2.enumerated().flatMap { [$1, lat.2[$0]] }
         // latlon - [1,1,0,1,0,0,0,1,0,0,1,0,1,0,1,1,0,1,1,1,1,1,0,1,0,1,1,1,1,...]
-        
+
         let bits = latlon.enumerated().reduce([String]()) { $1.0 % 5 > 0 ? $0 << $1.1 : $0 + $1.1 }
         //  bits: [11010,00100,10101,10111,11010,11110,01100,10110,10110,11011,10001,10010,10101,...]
-        
+
         let arr = bits.compactMap { charmap[$0] }
         // arr: [u,4,p,r,u,y,d,q,q,v,j,k,p,b,...]
-        
+
         return String(arr.prefix(length))
     }
-    
+
     // MARK: Private
 
     private static let bitmap = "0123456789bcdefghjkmnpqrstuvwxyz"
@@ -117,7 +117,7 @@ public extension Geohash {
         case sixtyCentimeters                   // ±0.00060 km
         case seventyFourMillimeters             // ±0.000074 km
     }
-    
+
     static func encode(latitude: Double, longitude: Double, precision: Precision) -> String {
         return encode(latitude: latitude, longitude: longitude, length: precision.rawValue)
     }
@@ -159,11 +159,11 @@ public extension CLLocationCoordinate2D {
             self = kCLLocationCoordinate2DInvalid
         }
     }
-    
+
     func geohash(length: Int) -> String {
         return Geohash.encode(latitude: latitude, longitude: longitude, length: length)
     }
-    
+
     func geohash(precision: Geohash.Precision) -> String {
         return geohash(length: precision.rawValue)
     }
@@ -171,12 +171,11 @@ public extension CLLocationCoordinate2D {
 
 #endif
 
-
 public extension Geohash {
     private static var base32 = "0123456789bcdefghjkmnpqrstuvwxyz"
     enum Direction: String {
         case n, e, s, w
-        
+
         var neighbor: [String] {
             switch self {
             case .n:
@@ -189,7 +188,7 @@ public extension Geohash {
                 return ["238967debc01fg45kmstqrwxuvhjyznp", "14365h7k9dcfesgujnmqp0r2twvyx8zb"]
             }
         }
-        
+
         var border: [String] {
             switch self {
             case .n:
@@ -207,10 +206,10 @@ public extension Geohash {
     static func adjacent(geohash: String, direction: Direction) throws -> String {
         let lastChar = geohash.last!
         var parent = String(geohash.dropLast())
-        let type = geohash.count % 2;
+        let type = geohash.count % 2
 
-        // Check for edge-cases which don"t share common prefix
-        if direction.border[type].contains(lastChar) && !parent.isEmpty {
+        // Check for edge-cases which don't share common prefix
+        if direction.border[type].contains(lastChar), !parent.isEmpty {
             parent = try Geohash.adjacent(geohash: parent, direction: direction)
         }
 
@@ -219,7 +218,7 @@ public extension Geohash {
 
         return parent + String(base32[charIndex])
     }
-    
+
     static func neighbors(geohash: String) throws -> [String] {
         let n = try adjacent(geohash: geohash, direction: .n)
         let e = try adjacent(geohash: geohash, direction: .e)
@@ -234,4 +233,21 @@ public extension Geohash {
             try adjacent(geohash: s, direction: .w) // sw
         ]
     }
+}
+
+private extension StringProtocol {
+    func distance(of element: Element) -> Int? { firstIndex(of: element)?.distance(in: self) }
+    func distance<S: StringProtocol>(of string: S) -> Int? { range(of: string)?.lowerBound.distance(in: self) }
+
+    subscript(offset: Int) -> Character {
+        self[index(startIndex, offsetBy: offset)]
+    }
+}
+
+private extension Collection {
+    func distance(to index: Index) -> Int { distance(from: startIndex, to: index) }
+}
+
+private extension String.Index {
+    func distance<S: StringProtocol>(in string: S) -> Int { string.distance(to: self) }
 }
